==> 4_exploratory_data_analysis_with_dtale.ipynb <==
# Exploratory Data Analysis with Dtale

This notebook utilizes the Dtale library for exploratory data analysis (EDA) on a Pandas DataFrame. It provides interactive features for visualizing, analyzing, and gaining insights into the dataset.

### Installs
The notebook installs the `dtale` library to enable interactive data analysis.

### Imports
Imports necessary libraries, including `dtale` and `pandas`, for data analysis.

### Configs
Configures settings for Dtale to work in the Colab environment.

### Execute
Loads the dataset into a Pandas DataFrame and sets up Dtale for interactive analysis. A link to the Dtale dashboard will be displayed below.

### Auxiliary Functions
Lists available instances of Dtale and provides a command to kill a specific instance if needed.

### Installs

!pip install -U dtale

### Imports

import dtale
import pandas as pd
import dtale.app as dtale_app

### Configs

dtale_app.USE_COLAB = True

### Execute

# Load the data into a Pandas DataFrame
df = pd.read_pickle(r"data2.pkl")

# Setup Dtale
# Link will be displayed below

table = pd.DataFrame(df)
d = dtale.show(table)
d

### Auxiliary Functions

#### Instances

dtale.instances()

#### ⚠️ Kill Instance ⚠️

# Kill an instance using this command
# uncomment and edit as needed...

# dtale.get_instance(1).kill()
